streamlit_speech2text-using-openai-whisper/requirements.txt
altair==6.0.0
attrs==25.4.0
audioop-lts==0.2.2
blinker==1.9.0
cachetools==6.2.5
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
filelock==3.20.3
fsspec==2026.1.0
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
llvmlite==0.46.0
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
narwhals==2.15.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
openai-whisper==20250625
packaging==26.0
pandas==2.3.3
pillow==12.1.0
protobuf==6.33.4
pyarrow==23.0.0
pydeck==0.9.1
pydub==0.25.1
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rpds-py==0.30.0
semantic-version==2.10.0
setuptools==80.10.2
setuptools-rust==1.12.0
six==1.17.0
smmap==5.0.2
SpeechRecognition==3.14.5
standard-aifc==3.13.0
standard-chunk==3.13.0
streamlit==1.53.1
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
toml==0.10.2
torch==2.10.0
tornado==6.5.4
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
